refactor: change type of size attribute in sector entity

+++ src/main/java/br/com/meli/projetointegrador/model/Sector.java
@@ -20,7 +20,7 @@ public class Sector {
     private Long id;
     private String name;
     private String category;
-    private Double size;
+    private Integer size;
 
     @ManyToOne
     private Warehouse warehouse;
